feat: 단일 상점 가게정보·원산지 조회 api

+++ src/main/java/in/koreatech/koin/domain/order/shop/model/domain/OrderableShopInfoSummary.java
@@ -8,7 +8,7 @@ public record OrderableShopInfoSummary(
     Boolean isTakeoutAvailable,
     Integer minimumOrderAmount,
     Double ratingAverage,
-    Long reviewCount,
+    Integer reviewCount,
     Integer minimumDeliveryTip,
     Integer maximumDeliveryTip
 ) {
